# api-contracts/openapi.yaml
openapi: 3.0.3
info:
  title: TrackBot App API
  description: |
    Comprehensive API for the TrackBot Application helping in logging workouts, exercises and nutrition using AI.
    
    This API provides endpoints for:
    - User management
    - Workout tracking and management
    - Exercise logging (cardio and weight training)
    - AI conversation and messaging
    
    All timestamps are in ISO 8601 format (UTC).
  version: 1.0.0
  contact:
    name: Utkarsh Khandelwal
    email: utkarshkhandelwal2011@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.trackbotapp.com/v1
    description: Production server
...
paths:
...
  /users:
    post:
      summary: Create a new user
      description: Create a new user account
      operationId: createUser
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/schemas/user.yaml#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: './components/schemas/user.yaml#/components/schemas/CreateUserResponse'
        '200':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: './components/schemas/user.yaml#/components/schemas/User'
        '400':
          $ref: './components/responses/errors.yaml#/components/responses/BadRequest'
        '409':
          $ref: './components/responses/errors.yaml#/components/responses/Conflict'
        '500':
          $ref: './components/responses/errors.yaml#/components/responses/InternalServerError'
  /users/email/{email}:
    get:
      summary: Get user by email
      description: Retrieve a user by their email address
      operationId: getUserByEmail
      tags: [Users]
      parameters:
        - $ref: './components/parameters/common.yaml#/EmailPath'
      responses:
        '200':
          description: User details retrieved successfully
          content:
            application/json:
              schema:
                $ref: './components/schemas/user.yaml#/components/schemas/User'
        '404':
          $ref: './components/responses/errors.yaml#/components/responses/NotFound'
        '500':
          $ref: './components/responses/errors.yaml#/components/responses/InternalServerError'

  # Workout endpoints
...
  /conversations/{conversationId}/messages:
    get:
      summary: List messages in a conversation
      description: Retrieve all messages in a specific conversation
      operationId: listMessages
      tags: [Messages]
      parameters:
        - $ref: './components/parameters/common.yaml#/ConversationIdPath'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of messages retrieved successfully
          content:
            application/json:
              schema:
                $ref: './components/schemas/message.yaml#/components/schemas/ListMessagesResponse'
        '404':
          $ref: './components/responses/errors.yaml#/components/responses/NotFound'
        '500':
          $ref: './components/responses/errors.yaml#/components/responses/InternalServerError'
    post:
      summary: Create a new message
      description: Send a new message in a conversation
      operationId: createMessage
      tags: [Messages]
      parameters:
        - $ref: './components/parameters/common.yaml#/ConversationIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/schemas/message.yaml#/components/schemas/CreateMessageRequest'
      responses:
        '201':
          description: Message created successfully
          content:
            application/json:
              schema:
                $ref: './components/schemas/message.yaml#/components/schemas/Message'
        '400':
          $ref: './components/responses/errors.yaml#/components/responses/BadRequest'
        '404':
          $ref: './components/responses/errors.yaml#/components/responses/NotFound'
        '500':
          $ref: './components/responses/errors.yaml#/components/responses/InternalServerError'
  /messages/{messageId}:
    get:
      summary: Get message by ID
      description: Retrieve a specific message by its ID
      operationId: getMessageById
      tags: [Messages]
      parameters:
        - $ref: './components/parameters/common.yaml#/MessageIdPath'
      responses:
        '200':
          description: Message details retrieved successfully
          content:
            application/json:
              schema:
                $ref: './components/schemas/message.yaml#/components/schemas/Message'
        '404':
          $ref: './components/responses/errors.yaml#/components/responses/NotFound'
        '500':
          $ref: './components/responses/errors.yaml#/components/responses/InternalServerError'
    delete:
      summary: Delete message
      description: Delete a message from a conversation
      operationId: deleteMessage
      tags: [Messages]
      parameters:
        - $ref: './components/parameters/common.yaml#/MessageIdPath'
      responses:
        '204':
          $ref: './components/responses/errors.yaml#/components/responses/NoContent'
        '404':
          $ref: './components/responses/errors.yaml#/components/responses/NotFound'
        '500':
          $ref: './components/responses/errors.yaml#/components/responses/InternalServerError'
components:
  schemas:
...
    User:
      $ref: './components/schemas/user.yaml#/components/schemas/User'
    CreateUserRequest:
      $ref: './components/schemas/user.yaml#/components/schemas/CreateUserRequest'
    CreateUserResponse:
      $ref: './components/schemas/user.yaml#/components/schemas/CreateUserResponse'
...
    Message:
      $ref: './components/schemas/message.yaml#/components/schemas/Message'
    CreateMessageRequest:
      $ref: './components/schemas/message.yaml#/components/schemas/CreateMessageRequest'
...
    ListMessagesResponse:
      $ref: './components/schemas/message.yaml#/components/schemas/ListMessagesResponse'
...
  responses:
    BadRequest:
      $ref: './components/responses/errors.yaml#/components/responses/BadRequest'
...
    NotFound:
      $ref: './components/responses/errors.yaml#/components/responses/NotFound'
    Conflict:
      $ref: './components/responses/errors.yaml#/components/responses/Conflict'
...
    InternalServerError:
      $ref: './components/responses/errors.yaml#/components/responses/InternalServerError'
    NoContent:
      $ref: './components/responses/errors.yaml#/components/responses/NoContent'